gemini/workers/ml/requirements.txt
minio>=7.2,<8
redis>=5,<6
requests>=2.31,<3
Pillow>=10,<11
numpy>=1.24,<2
rasterio>=1.3,<2
geopandas>=0.14,<1.1
opencv-python-headless>=4.8,<5
